Rely on scrollIntoView() when doing feature detection of RTL behavior
The Chromium community is moving to standard behavior from the CSSOM
specification [1]. In order to launch the feature and analyze potential
broken URLs, they are counting pages setting the scrollLeft of a RTL
scroller to a positive value. Websites using OverlayScrollbars appear
in the top results and cause a lot of false positives [2]. In order to
workaround that issue, rely on Element.scrollIntoView() instead.

[1] https://www.chromestatus.com/feature/5759578031521792
[2] https://github.com/KingSora/OverlayScrollbars/issues/187